+++ src/vue/forms/market-orders/SubmitTradeOfferForm.vue
@@ -1,24 +1,38 @@
 <template>
-  <form>
+  <form
+    v-if="isLoaded"
+    @submit.prevent="isFormValid() && showConfirmation()"
+  >
+    <div class="app__form-row">
+      <div class="app__form-field">
+        <div class="app__form-field">
+          <readonly-field
+            :label="baseAssetLabelTranslationId | globalize"
+            :value="offer.baseAsset.id"
+          />
+        </div>
+      </div>
+    </div>
+
     <div class="app__form-row">
       <div class="app__form-field">
         <input-field
           v-model.trim="form.price"
-          name="submit-trade-offers-offer-price"
+          name="submit-trade-offer-price"
           :label="
-            'submit-trade-offers-form.price-label' | globalize({
-              asset: assetPair.quote
+            'submit-trade-offer-form.price-lbl' | globalize({
+              baseAsset: assetPair.base,
+              quoteAsset: assetPair.quote,
             })
           "
           :error-message="getFieldErrorMessage(
             'form.price', {
               from: config.MIN_AMOUNT,
               to: config.MAX_AMOUNT,
-              available: quoteAssetBalance.balance
+              available: quoteAssetBalance
             }
           )"
           @blur="touchField('form.price')"
-          :white-autofill="true"
           :disabled="formMixin.isDisabled"
         />
       </div>
@@ -28,22 +42,19 @@
       <div class="app__form-field">
         <input-field
           v-model.trim="form.baseAmount"
-          name="submit-trade-offers-offer-base-amount"
-          :label="'submit-trade-offers-form.want-label' | globalize({
+          name="submit-trade-offer-base-amount"
+          :label="'submit-trade-offer-form.base-amount-lbl' | globalize({
             asset: assetPair.base
           })"
           :error-message="getFieldErrorMessage(
             'form.baseAmount',
             {
-              minValue: config.MIN_AMOUNT,
-              available: baseAssetBalance.balance,
+              available: baseAssetBalance,
               from: config.MIN_AMOUNT,
               to: config.MAX_AMOUNT,
             }
           )"
-          @blur="touchField('form.amount')"
-          :white-autofill="true"
-          :max="baseAssetBalance.balance"
+          @blur="touchField('form.baseAmount')"
           :disabled="formMixin.isDisabled"
         />
       </div>
@@ -51,195 +62,105 @@
 
     <div class="app__form-row">
       <div class="app__form-field">
-        <input-field
+        <readonly-field
           :label="
-            'submit-trade-offers-form.offer-label' | globalize({
+            'submit-trade-offer-form.total-amount-lbl' | globalize({
               asset: assetPair.quote
             })
           "
-          v-model.trim="totalValue"
-          name="submit-trade-offers-offer-quote-amount"
-          :white-autofill="true"
+          :value="{
+            value: quoteAmount,
+            currency: assetPair.quote,
+          } | formatMoney"
           :error-message="getFieldErrorMessage(
-            'totalValue',
+            'quoteAmount',
             {
-              minValue: config.MIN_AMOUNT,
               available: isBuy ? quoteAssetBalance : baseAssetBalance,
               from: config.MIN_AMOUNT,
               to: config.MAX_AMOUNT,
             }
           )"
-          @change="touchField('totalValue')"
-          :readonly="true"
         />
       </div>
     </div>
 
     <template v-if="formMixin.isConfirmationShown">
       <form-confirmation
-        @cancel="hideConfirmation"
-        @ok="submit"
-        :is-pending="isOfferSubmitting"
         class="app__form-confirmation"
+        :is-pending="isOfferSubmitting"
+        @ok="submit"
+        @cancel="hideConfirmation"
       />
     </template>
+
     <template v-else>
-      <div class="submit-trade-offer-form__actions">
-        <!-- TODO: make it via tooltip message -->
-        <p
-          v-if="!isEnoughOnBalance && !isOwner"
-          class="app__form-field-description"
+      <div class="app__form-actions">
+        <button
+          v-ripple
+          class="app__button-raised submit-trade-offer-form__btn"
+          type="submit"
+          :disabled="formMixin.isDisabled"
         >
-          {{
-            'submit-trade-offers-form.insufficient-funds' | globalize({
-              amount: formatNumber(isBuy
-                ? baseAssetBalance.balance
-                : quoteAssetBalance.balance)
-            })
-          }}
-        </p>
-        <div class="app__form-actions">
-          <button
-            v-ripple
-            v-if="isOwner"
-            type="button"
-            @click="showConfirmation"
-            class="app__form-submit-btn app__button-raised"
-            :disabled="formMixin.isDisabled"
-          >
-            {{ 'submit-trade-offers-form.cancel-offer-btn' | globalize }}
-          </button>
-          <button
-            v-ripple
-            v-else
-            type="button"
-            @click="showConfirmation"
-            class="app__form-submit-btn app__button-raised"
-            :disabled="isDisabledSubmitButton"
-          >
-            <template v-if="isBuy">
-              {{ 'submit-trade-offers-form.submit-sell-btn' | globalize }}
-            </template>
-            <template v-else>
-              {{ 'submit-trade-offers-form.submit-buy-btn' | globalize }}
-            </template>
-          </button>
-        </div>
+          <template v-if="isBuy">
+            {{ 'submit-trade-offer-form.buy-btn' | globalize }}
+          </template>
+
+          <template v-else>
+            {{ 'submit-trade-offer-form.sell-btn' | globalize }}
+          </template>
+        </button>
       </div>
     </template>
   </form>
 </template>
 
 <script>
+import ReadonlyField from '@/vue/fields/ReadonlyField'
+
 import FormMixin from '@/vue/mixins/form.mixin'
 import OfferManagerMixin from '@/vue/mixins/offer-manager.mixin'
-import FormConfirmation from '@/vue/common/FormConfirmation'
-import { formatNumber } from '@/vue/filters/formatNumber'
-import { vuexTypes } from '@/vuex'
-import { mapGetters, mapActions } from 'vuex'
+
+import { Bus } from '@/js/helpers/event-bus'
+import { ErrorHandler } from '@/js/helpers/error-handler'
+
+import { MathUtil } from '@/js/utils/math.util'
 import config from '@/config'
+
 import {
   required,
   amountRange,
-  minValue,
   noMoreThanAvailableOnBalance,
   decimal,
 } from '@validators'
-import { MathUtil } from '@/js/utils/math.util'
 
 const EVENTS = {
-  closeDrawer: 'close-drawer',
+  offerSubmitted: 'offer-submitted',
 }
 
 export default {
   name: 'submit-trade-offer-form',
-  components: {
-    FormConfirmation,
-  },
+  components: { ReadonlyField },
   mixins: [
     FormMixin,
     OfferManagerMixin,
   ],
+
   props: {
-    assetPair: {
-      type: Object,
-      default: () => ({}),
-    },
-    isBuy: {
-      type: Boolean,
-      default: false,
-    },
-    offer: {
-      type: Object,
-      require: true,
-      default: () => ({}),
-    },
+    assetPair: { type: Object, required: true },
+    isBuy: { type: Boolean, default: false },
+    offer: { type: Object, required: true },
   },
+
   data: () => ({
-    config,
-    isOfferSubmitting: false,
-    isCancelingOrder: false,
     form: {
-      form: {
-        baseAmount: '0',
-        baseAssetCode: null,
-        offerId: null,
-        ownerId: null,
-        price: '0',
-        quoteAmount: null,
-        quoteAssetCode: null,
-        quoteBalanceId: null,
-      },
+      price: '',
+      baseAmount: '',
     },
+    isLoaded: false,
+    isOfferSubmitting: false,
+    config,
   }),
-  computed: {
-    ...mapGetters([
-      vuexTypes.accountId,
-      vuexTypes.accountBalances,
-    ]),
-    baseAssetBalance () {
-      return this.accountBalances
-        .find(item => item.asset === this.form.baseAssetCode) || {}
-    },
-    quoteAssetBalance () {
-      return this.accountBalances
-        .find(item => item.asset === this.form.quoteAssetCode) || {}
-    },
-    isEnoughOnBalance () {
-      if (this.isBuy) {
-        return +this.baseAssetBalance.balance >= +this.form.baseAmount
-      }
-      return +this.quoteAssetBalance.balance >= +this.form.baseAmount
-    },
-    isDisabledSubmitButton () {
-      return !this.isEnoughOnBalance ||
-        this.formMixin.isDisabled ||
-        !this.form.price ||
-        !this.form.baseAmount
-    },
-    formQuoteAmount () {
-      return MathUtil.multiply(this.form.price, this.form.baseAmount)
-    },
-    totalValue () {
-      return +this.formQuoteAmount ? this.formQuoteAmount : ''
-    },
-    isOwner () {
-      return this.form.ownerId === this.accountId
-    },
-  },
-  async created () {
-    await this.loadBalances()
-    this.form = {
-      baseAmount: this.offer.baseAmount,
-      baseAssetCode: this.offer.baseAssetCode,
-      offerId: this.offer.offerId,
-      ownerId: this.offer.ownerId,
-      price: this.offer.price,
-      quoteAmount: this.offer.quoteAmount,
-      quoteAssetCode: this.offer.quoteAssetCode,
-      quoteBalanceId: this.offer.quoteBalanceId,
-    }
-  },
+
   validations () {
     return {
       form: {
@@ -251,78 +172,99 @@ export default {
         baseAmount: {
           required,
           decimal,
-          minValue: minValue(config.MIN_AMOUNT),
-          noMoreThanAvailableOnBalance: this.isBuy
-            ? this.isBuy
-            : noMoreThanAvailableOnBalance(this.baseAssetBalance),
+          noMoreThanAvailableOnBalance: this.isBuy ||
+            noMoreThanAvailableOnBalance(this.baseAssetBalance),
           amountRange: amountRange(config.MIN_AMOUNT, config.MAX_AMOUNT),
         },
       },
-      totalValue: {
-        noMoreThanAvailableOnBalance: this.isBuy
-          ? noMoreThanAvailableOnBalance(this.quoteAssetBalance)
-          : !this.isBuy,
+
+      quoteAmount: {
+        noMoreThanAvailableOnBalance: !this.isBuy ||
+          noMoreThanAvailableOnBalance(this.quoteAssetBalance),
         amountRange: amountRange(config.MIN_AMOUNT, config.MAX_AMOUNT),
       },
     }
   },
-  methods: {
-    ...mapActions({
-      loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
-    }),
-    formatNumber,
-    async submit () {
-      this.disableForm()
-      this.isOfferSubmitting = true
 
-      if (this.isOwner) {
-        await this.cancelOffer(
-          this.getCancelOfferOpts()
-        )
-      } else {
-        await this.createOffer(
-          this.getCreateOfferOpts()
-        )
-      }
-
-      this.isOfferSubmitting = false
-      this.enableForm()
-      this.$emit(EVENTS.closeDrawer)
-      this.hideConfirmation()
+  computed: {
+    baseAssetLabelTranslationId () {
+      return this.isBuy
+        ? 'submit-trade-offer-form.asset-to-buy-lbl'
+        : 'submit-trade-offer-form.asset-to-sell-lbl'
     },
-    getCreateOfferOpts () {
+
+    baseAssetBalance () {
+      const balanceItem = this.accountBalances
+        .find(balance => balance.asset === this.assetPair.base)
+      return balanceItem ? balanceItem.balance : ''
+    },
+
+    quoteAssetBalance () {
+      const balanceItem = this.accountBalances
+        .find(balance => balance.asset === this.assetPair.quote)
+      return balanceItem ? balanceItem.balance : ''
+    },
+
+    quoteAmount () {
+      if (this.form.price && this.form.baseAmount) {
+        return MathUtil.multiply(this.form.price, this.form.baseAmount)
+      } else {
+        return ''
+      }
+    },
+
+    createOfferOpts () {
       return {
         pair: {
-          base: this.form.baseAssetCode,
-          quote: this.form.quoteAssetCode,
+          base: this.assetPair.base,
+          quote: this.assetPair.quote,
         },
         baseAmount: this.form.baseAmount,
-        quoteAmount: this.form.quoteAmount,
+        quoteAmount: this.quoteAmount,
         price: this.form.price,
-        isBuy: !this.form.isBuy,
+        isBuy: this.isBuy,
       }
     },
-    getCancelOfferOpts () {
-      return {
-        price: this.form.price,
-        baseBalance: this.baseAssetBalance.balanceId,
-        quoteBalance: this.quoteAssetBalance.balanceId,
-        offerId: this.form.offerId,
+  },
+
+  async created () {
+    try {
+      await this.loadBalances()
+      this.populateForm()
+      this.isLoaded = true
+    } catch (e) {
+      ErrorHandler.processWithoutFeedback(e)
+    }
+  },
+
+  methods: {
+    populateForm () {
+      this.form.price = this.offer.price
+      this.form.baseAmount = this.offer.baseAmount
+    },
+
+    async submit () {
+      this.isOfferSubmitting = true
+      try {
+        await this.createOffer(this.createOfferOpts)
+
+        Bus.success('submit-trade-offer-form.order-submitted-msg')
+        this.$emit(EVENTS.offerSubmitted)
+      } catch (e) {
+        ErrorHandler.process(e)
       }
+      this.isOfferSubmitting = false
+      this.hideConfirmation()
     },
   },
 }
 </script>
 
 <style lang="scss" scoped>
-  @import '../app-form';
+@import '../app-form';
 
-  .submit-trade-offer-form__actions {
-    margin-top: 5rem;
-    /* stylelint-disable selector-nested-pattern */
-    & > .app__form-actions {
-      margin-top: 1.6rem;
-    }
-    /* stylelint-enable selector-nested-pattern */
-  }
+.submit-trade-offer-form__btn {
+  max-width: 14rem;
+  width: 100%;
+}
 </style>