+++ src/vue/modules/create-sale-form/index.vue
@@ -71,8 +71,6 @@ import { ErrorHandler } from '@/js/helpers/error-handler'
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
-import { initConfig } from './_config'
-
 const STEPS = {
   information: {
     number: 1,
@@ -104,10 +102,6 @@ export default {
   },
   mixins: [LoadAssetsMixin, ManageSaleRequestMixin],
   props: {
-    storageUrl: {
-      type: String,
-      required: true,
-    },
     requestId: {
       type: String,
       default: '',
@@ -140,8 +134,6 @@ export default {
   methods: {
     async init () {
       try {
-        initConfig(this.storageUrl)
-
         await this.loadAssets(this.accountId)
         await this.tryLoadRequest()